feat(fileprovider): add daemon-backed gRPC backend (feature-gated)
Refactor tcfs-file-provider into two feature-gated backends:

- `direct` (default): Original S3/OpenDAL backend. No daemon needed
  but no fleet sync or NATS events. Zero behavioral change.

- `grpc`: New backend that delegates all operations to tcfsd daemon
  via Unix domain socket gRPC. Enables full fleet sync, NATS event
  propagation, and conflict resolution through the daemon.

  Config accepts `daemon_socket` field (falls back to $TCFS_SOCKET
  or $XDG_STATE_HOME/tcfsd/tcfsd.sock).

  Maps FFI calls to daemon RPCs:
  - enumerate → SyncStatus
  - fetch → Hydrate (streaming)
  - upload → Push (streaming)
  - delete → Unsync
  - create_dir → Push (zero-byte marker)

Shared code (TcfsError, TcfsFileItem, memory management) stays in
lib.rs. Each backend is a separate module with identical C ABI.
