+++ crates/tcfs-file-provider/Cargo.toml
@@ -9,20 +9,31 @@ description = "tcfs C FFI bridge for macOS/iOS FileProvider extensions via cbind
 [lib]
 crate-type = ["lib", "staticlib"]
 
+[features]
+default = ["direct"]
+## Direct S3/OpenDAL backend (current, no daemon needed)
+direct = ["dep:tcfs-storage", "dep:tcfs-chunks", "dep:tcfs-crypto", "dep:opendal", "dep:secrecy", "dep:base64"]
+## gRPC backend — delegates all operations to tcfsd daemon
+grpc = ["dep:tonic", "dep:tower", "dep:hyper-util", "dep:tokio-stream"]
+
 [dependencies]
 tcfs-core = { path = "../tcfs-core" }
-tcfs-storage = { path = "../tcfs-storage" }
-tcfs-chunks = { path = "../tcfs-chunks" }
-tcfs-crypto = { path = "../tcfs-crypto" }
+tcfs-storage = { path = "../tcfs-storage", optional = true }
+tcfs-chunks = { path = "../tcfs-chunks", optional = true }
+tcfs-crypto = { path = "../tcfs-crypto", optional = true }
 tcfs-sync = { path = "../tcfs-sync" }
-secrecy = { workspace = true }
-base64 = { workspace = true }
-opendal = { workspace = true }
+secrecy = { workspace = true, optional = true }
+base64 = { workspace = true, optional = true }
+opendal = { workspace = true, optional = true }
 tokio = { workspace = true }
 anyhow = { workspace = true }
 thiserror = { workspace = true }
 tracing = { workspace = true }
 serde_json = { workspace = true }
+tonic = { workspace = true, optional = true }
+tower = { workspace = true, optional = true }
+hyper-util = { workspace = true, optional = true }
+tokio-stream = { workspace = true, optional = true }
 
 [build-dependencies]
 cbindgen = "0.27"
