+++ crates/tcfs-file-provider/src/grpc_backend.rs
@@ -1,0 +1,508 @@
+//! gRPC backend for the FileProvider FFI.
+//!
+//! Delegates all operations to the tcfsd daemon via Unix domain socket gRPC.
+//! This enables full fleet sync, NATS events, and conflict resolution —
+//! the daemon handles E2EE, chunking, and storage internally.
+
+use std::ffi::CStr;
+use std::os::raw::c_char;
+use std::panic::AssertUnwindSafe;
+use std::path::PathBuf;
+use std::ptr;
+
+use tcfs_core::proto::tcfs_daemon_client::TcfsDaemonClient;
+use tonic::transport::{Channel, Endpoint, Uri};
+use tower::service_fn;
+
+use crate::{to_c_string, TcfsError, TcfsFileItem};
+
+/// Opaque provider handle wrapping a tokio runtime + gRPC client.
+///
+/// Created via `tcfs_provider_new`, freed via `tcfs_provider_free`.
+pub struct TcfsProvider {
+    runtime: tokio::runtime::Runtime,
+    client: TcfsDaemonClient<Channel>,
+    /// Remote prefix for path construction
+    remote_prefix: String,
+}
+
+/// Connect to the daemon over a Unix domain socket.
+async fn connect(socket_path: &str) -> Result<TcfsDaemonClient<Channel>, anyhow::Error> {
+    let path = PathBuf::from(socket_path);
+
+    let channel = Endpoint::from_static("http://[::]:0")
+        .connect_with_connector(service_fn(move |_: Uri| {
+            let path = path.clone();
+            async move {
+                let stream = tokio::net::UnixStream::connect(&path).await?;
+                Ok::<_, std::io::Error>(hyper_util::rt::TokioIo::new(stream))
+            }
+        }))
+        .await?;
+
+    Ok(TcfsDaemonClient::new(channel))
+}
+
+/// Create a new provider from a JSON configuration string.
+///
+/// The JSON should contain:
+/// ```json
+/// {
+///   "daemon_socket": "/path/to/tcfsd.sock",
+///   "remote_prefix": "devices/mydevice"
+/// }
+/// ```
+///
+/// Falls back to `$TCFS_SOCKET` env var, then
+/// `$XDG_STATE_HOME/tcfsd/tcfsd.sock` if `daemon_socket` is not set.
+///
+/// # Safety
+///
+/// `config_json` must be a valid null-terminated UTF-8 C string.
+#[no_mangle]
+pub unsafe extern "C" fn tcfs_provider_new(config_json: *const c_char) -> *mut TcfsProvider {
+    if config_json.is_null() {
+        return ptr::null_mut();
+    }
+
+    let result = std::panic::catch_unwind(AssertUnwindSafe(|| {
+        let c_str = unsafe { CStr::from_ptr(config_json) };
+        let json_str = match c_str.to_str() {
+            Ok(s) => s,
+            Err(_) => return ptr::null_mut(),
+        };
+
+        let config: serde_json::Value = match serde_json::from_str(json_str) {
+            Ok(v) => v,
+            Err(_) => return ptr::null_mut(),
+        };
+
+        let socket_path = config["daemon_socket"]
+            .as_str()
+            .map(|s| s.to_string())
+            .or_else(|| std::env::var("TCFS_SOCKET").ok())
+            .unwrap_or_else(|| {
+                let state_home = std::env::var("XDG_STATE_HOME").unwrap_or_else(|_| {
+                    let home = std::env::var("HOME").unwrap_or_else(|_| "/tmp".into());
+                    format!("{home}/.local/state")
+                });
+                format!("{state_home}/tcfsd/tcfsd.sock")
+            });
+
+        let prefix = config["remote_prefix"]
+            .as_str()
+            .unwrap_or("default")
+            .to_string();
+
+        let runtime = match tokio::runtime::Runtime::new() {
+            Ok(rt) => rt,
+            Err(_) => return ptr::null_mut(),
+        };
+
+        let client = match runtime.block_on(connect(&socket_path)) {
+            Ok(c) => c,
+            Err(e) => {
+                tracing::error!("failed to connect to tcfsd at {}: {}", socket_path, e);
+                return ptr::null_mut();
+            }
+        };
+
+        Box::into_raw(Box::new(TcfsProvider {
+            runtime,
+            client,
+            remote_prefix: prefix,
+        }))
+    }));
+
+    result.unwrap_or(ptr::null_mut())
+}
+
+/// Enumerate files by querying daemon sync status.
+///
+/// # Safety
+///
+/// - `provider` must be a valid pointer from `tcfs_provider_new`.
+/// - `path` must be a valid null-terminated UTF-8 C string (use "" for root).
+/// - `out_items` and `out_count` must be valid writable pointers.
+#[no_mangle]
+pub unsafe extern "C" fn tcfs_provider_enumerate(
+    provider: *mut TcfsProvider,
+    path: *const c_char,
+    out_items: *mut *mut TcfsFileItem,
+    out_count: *mut usize,
+) -> TcfsError {
+    if provider.is_null() || path.is_null() || out_items.is_null() || out_count.is_null() {
+        return TcfsError::TcfsErrorInvalidArg;
+    }
+
+    let result = std::panic::catch_unwind(AssertUnwindSafe(|| {
+        let prov = unsafe { &mut *provider };
+        let c_path = unsafe { CStr::from_ptr(path) };
+        let rel_path = match c_path.to_str() {
+            Ok(s) => s,
+            Err(_) => return TcfsError::TcfsErrorInvalidArg,
+        };
+
+        let enumerate_result = prov.runtime.block_on(async {
+            // Query sync status for the path — the daemon returns file metadata
+            let resp = prov
+                .client
+                .sync_status(tonic::Request::new(
+                    tcfs_core::proto::SyncStatusRequest {
+                        path: rel_path.to_string(),
+                    },
+                ))
+                .await?;
+
+            let status = resp.into_inner();
+
+            // The daemon returns a single file's status. For directory enumeration,
+            // we need to list files. Use the Watch RPC to get initial state, or
+            // fall back to status queries. For now, return the single item if found.
+            let mut items: Vec<TcfsFileItem> = Vec::new();
+
+            if !status.path.is_empty() && status.state != "not_found" {
+                let filename = status
+                    .path
+                    .rsplit('/')
+                    .next()
+                    .unwrap_or(&status.path);
+                let is_dir = status.state == "directory";
+
+                items.push(TcfsFileItem {
+                    item_id: to_c_string(&status.path),
+                    filename: to_c_string(filename),
+                    file_size: status.size,
+                    modified_timestamp: status.last_synced,
+                    is_directory: is_dir,
+                    content_hash: to_c_string(&status.blake3),
+                });
+            }
+
+            Ok::<Vec<TcfsFileItem>, tonic::Status>(items)
+        });
+
+        match enumerate_result {
+            Ok(items) => {
+                let count = items.len();
+                let boxed = items.into_boxed_slice();
+                let ptr = Box::into_raw(boxed) as *mut TcfsFileItem;
+                unsafe {
+                    *out_items = ptr;
+                    *out_count = count;
+                }
+                TcfsError::TcfsErrorNone
+            }
+            Err(e) => {
+                tracing::error!("enumerate failed: {}", e);
+                TcfsError::TcfsErrorStorage
+            }
+        }
+    }));
+
+    result.unwrap_or(TcfsError::TcfsErrorInternal)
+}
+
+/// Fetch (hydrate) a file via the daemon's Hydrate RPC.
+///
+/// # Safety
+///
+/// - `provider` must be a valid pointer from `tcfs_provider_new`.
+/// - `item_id` and `dest_path` must be valid null-terminated UTF-8 C strings.
+#[no_mangle]
+pub unsafe extern "C" fn tcfs_provider_fetch(
+    provider: *mut TcfsProvider,
+    item_id: *const c_char,
+    dest_path: *const c_char,
+) -> TcfsError {
+    if provider.is_null() || item_id.is_null() || dest_path.is_null() {
+        return TcfsError::TcfsErrorInvalidArg;
+    }
+
+    let result = std::panic::catch_unwind(AssertUnwindSafe(|| {
+        let prov = unsafe { &mut *provider };
+        let c_item = unsafe { CStr::from_ptr(item_id) };
+        let c_dest = unsafe { CStr::from_ptr(dest_path) };
+
+        let item_str = match c_item.to_str() {
+            Ok(s) => s,
+            Err(_) => return TcfsError::TcfsErrorInvalidArg,
+        };
+        let dest_str = match c_dest.to_str() {
+            Ok(s) => s,
+            Err(_) => return TcfsError::TcfsErrorInvalidArg,
+        };
+
+        let fetch_result = prov.runtime.block_on(async {
+            // Use Hydrate RPC — daemon handles download, decryption, assembly
+            let mut stream = prov
+                .client
+                .hydrate(tonic::Request::new(
+                    tcfs_core::proto::HydrateRequest {
+                        stub_path: item_str.to_string(),
+                        partial_ok: false,
+                    },
+                ))
+                .await?
+                .into_inner();
+
+            let mut local_path = String::new();
+            while let Some(progress) = stream.message().await? {
+                if !progress.error.is_empty() {
+                    return Err(tonic::Status::internal(progress.error));
+                }
+                if progress.done {
+                    local_path = progress.local_path;
+                    break;
+                }
+            }
+
+            if local_path.is_empty() {
+                return Err(tonic::Status::internal("hydrate completed without local_path"));
+            }
+
+            // Copy the daemon's hydrated file to the FileProvider's destination
+            tokio::fs::copy(&local_path, dest_str)
+                .await
+                .map_err(|e| tonic::Status::internal(format!("copy to dest: {e}")))?;
+
+            Ok::<(), tonic::Status>(())
+        });
+
+        match fetch_result {
+            Ok(()) => TcfsError::TcfsErrorNone,
+            Err(e) => {
+                tracing::error!("fetch failed: {}", e);
+                TcfsError::TcfsErrorStorage
+            }
+        }
+    }));
+
+    result.unwrap_or(TcfsError::TcfsErrorInternal)
+}
+
+/// Upload a local file via the daemon's Push RPC (streaming).
+///
+/// # Safety
+///
+/// - `provider` must be a valid pointer from `tcfs_provider_new`.
+/// - `local_path` and `remote_rel` must be valid null-terminated UTF-8 C strings.
+#[no_mangle]
+pub unsafe extern "C" fn tcfs_provider_upload(
+    provider: *mut TcfsProvider,
+    local_path: *const c_char,
+    remote_rel: *const c_char,
+) -> TcfsError {
+    if provider.is_null() || local_path.is_null() || remote_rel.is_null() {
+        return TcfsError::TcfsErrorInvalidArg;
+    }
+
+    let result = std::panic::catch_unwind(AssertUnwindSafe(|| {
+        let prov = unsafe { &mut *provider };
+        let c_local = unsafe { CStr::from_ptr(local_path) };
+        let c_remote = unsafe { CStr::from_ptr(remote_rel) };
+
+        let local_str = match c_local.to_str() {
+            Ok(s) => s,
+            Err(_) => return TcfsError::TcfsErrorInvalidArg,
+        };
+        let remote_str = match c_remote.to_str() {
+            Ok(s) => s,
+            Err(_) => return TcfsError::TcfsErrorInvalidArg,
+        };
+
+        let upload_result = prov.runtime.block_on(async {
+            let data = tokio::fs::read(local_str).await.map_err(|e| {
+                tonic::Status::internal(format!("read local file: {e}"))
+            })?;
+
+            let remote_path = format!(
+                "{}/{}",
+                prov.remote_prefix.trim_end_matches('/'),
+                remote_str.trim_start_matches('/')
+            );
+
+            // Send the file as a single PushChunk (daemon handles chunking internally)
+            let chunk = tcfs_core::proto::PushChunk {
+                path: remote_path,
+                data: data.clone(),
+                offset: 0,
+                last: true,
+            };
+
+            let stream = tokio_stream::once(chunk);
+            let mut resp_stream = prov
+                .client
+                .push(tonic::Request::new(stream))
+                .await?
+                .into_inner();
+
+            // Drain progress stream
+            while let Some(progress) = resp_stream.message().await? {
+                if !progress.error.is_empty() {
+                    return Err(tonic::Status::internal(progress.error));
+                }
+                if progress.done {
+                    break;
+                }
+            }
+
+            Ok::<(), tonic::Status>(())
+        });
+
+        match upload_result {
+            Ok(()) => TcfsError::TcfsErrorNone,
+            Err(e) => {
+                tracing::error!("upload failed: {}", e);
+                TcfsError::TcfsErrorStorage
+            }
+        }
+    }));
+
+    result.unwrap_or(TcfsError::TcfsErrorInternal)
+}
+
+/// Delete a file via the daemon (unsync then delete remote).
+///
+/// # Safety
+///
+/// - `provider` must be a valid pointer from `tcfs_provider_new`.
+/// - `item_id` must be a valid null-terminated UTF-8 C string.
+#[no_mangle]
+pub unsafe extern "C" fn tcfs_provider_delete(
+    provider: *mut TcfsProvider,
+    item_id: *const c_char,
+) -> TcfsError {
+    if provider.is_null() || item_id.is_null() {
+        return TcfsError::TcfsErrorInvalidArg;
+    }
+
+    let result = std::panic::catch_unwind(AssertUnwindSafe(|| {
+        let prov = unsafe { &mut *provider };
+        let c_item = unsafe { CStr::from_ptr(item_id) };
+        let item_str = match c_item.to_str() {
+            Ok(s) => s,
+            Err(_) => return TcfsError::TcfsErrorInvalidArg,
+        };
+
+        let delete_result = prov.runtime.block_on(async {
+            let resp = prov
+                .client
+                .unsync(tonic::Request::new(
+                    tcfs_core::proto::UnsyncRequest {
+                        path: item_str.to_string(),
+                        force: true,
+                    },
+                ))
+                .await?
+                .into_inner();
+
+            if !resp.success && !resp.error.is_empty() {
+                return Err(tonic::Status::internal(resp.error));
+            }
+
+            Ok::<(), tonic::Status>(())
+        });
+
+        match delete_result {
+            Ok(()) => TcfsError::TcfsErrorNone,
+            Err(e) => {
+                tracing::error!("delete failed: {}", e);
+                TcfsError::TcfsErrorStorage
+            }
+        }
+    }));
+
+    result.unwrap_or(TcfsError::TcfsErrorInternal)
+}
+
+/// Create a directory via the daemon's Push RPC (zero-byte marker).
+///
+/// # Safety
+///
+/// - `provider` must be a valid pointer from `tcfs_provider_new`.
+/// - `parent_path` and `dir_name` must be valid null-terminated UTF-8 C strings.
+#[no_mangle]
+pub unsafe extern "C" fn tcfs_provider_create_dir(
+    provider: *mut TcfsProvider,
+    parent_path: *const c_char,
+    dir_name: *const c_char,
+) -> TcfsError {
+    if provider.is_null() || parent_path.is_null() || dir_name.is_null() {
+        return TcfsError::TcfsErrorInvalidArg;
+    }
+
+    let result = std::panic::catch_unwind(AssertUnwindSafe(|| {
+        let prov = unsafe { &mut *provider };
+        let c_parent = unsafe { CStr::from_ptr(parent_path) };
+        let c_name = unsafe { CStr::from_ptr(dir_name) };
+
+        let parent_str = match c_parent.to_str() {
+            Ok(s) => s,
+            Err(_) => return TcfsError::TcfsErrorInvalidArg,
+        };
+        let name_str = match c_name.to_str() {
+            Ok(s) => s,
+            Err(_) => return TcfsError::TcfsErrorInvalidArg,
+        };
+
+        let dir_path = format!(
+            "{}/{}{}/",
+            prov.remote_prefix.trim_end_matches('/'),
+            if parent_str.is_empty() {
+                String::new()
+            } else {
+                format!("{}/", parent_str.trim_matches('/'))
+            },
+            name_str.trim_matches('/')
+        );
+
+        let create_result = prov.runtime.block_on(async {
+            let chunk = tcfs_core::proto::PushChunk {
+                path: dir_path,
+                data: vec![],
+                offset: 0,
+                last: true,
+            };
+
+            let stream = tokio_stream::once(chunk);
+            let mut resp_stream = prov
+                .client
+                .push(tonic::Request::new(stream))
+                .await?
+                .into_inner();
+
+            while let Some(progress) = resp_stream.message().await? {
+                if progress.done {
+                    break;
+                }
+            }
+
+            Ok::<(), tonic::Status>(())
+        });
+
+        match create_result {
+            Ok(()) => TcfsError::TcfsErrorNone,
+            Err(e) => {
+                tracing::error!("create_dir failed: {}", e);
+                TcfsError::TcfsErrorStorage
+            }
+        }
+    }));
+
+    result.unwrap_or(TcfsError::TcfsErrorInternal)
+}
+
+/// Free a provider handle.
+///
+/// # Safety
+///
+/// `provider` must be a valid pointer from `tcfs_provider_new`, or null.
+#[no_mangle]
+pub unsafe extern "C" fn tcfs_provider_free(provider: *mut TcfsProvider) {
+    if !provider.is_null() {
+        unsafe {
+            drop(Box::from_raw(provider));
+        }
+    }
+}
