+++ crates/tcfs-file-provider/src/lib.rs
@@ -4,35 +4,15 @@
 //! as a C-compatible FFI layer via cbindgen, enabling Swift consumers
 //! to build native FileProvider extensions (.appex).
 //!
-//! ## Architecture
+//! ## Backends
 //!
-//! ```text
-//! iOS Files App / macOS Finder
-//!       |
-//!       +-- NSFileProviderReplicatedExtension (Swift ~200 LOC)
-//!       |         |
-//!       |         +-- C header (tcfs_file_provider.h, cbindgen-generated)
-//!       |                   |
-//!       +-- tcfs-file-provider (this crate, staticlib)
-//!                   |
-//!                   +-- tcfs-storage  -> S3/SeaweedFS access
-//!                   +-- tcfs-chunks   -> FastCDC + BLAKE3 + zstd
-//!                   +-- tcfs-sync     -> state cache, manifests
-//!                   +-- tcfs-core     -> config, proto types
-//! ```
-//!
-//! ## Status
-//!
-//! Initial FFI skeleton -- see [RFC 0002](../../docs/rfc/0002-darwin-integration.md)
-//! and [RFC 0003](../../docs/rfc/0003-ios-file-provider.md) for roadmap.
+//! - `direct` (default): Talks directly to S3/SeaweedFS via OpenDAL.
+//!   No daemon needed but no fleet sync, no NATS events.
+//! - `grpc`: Delegates all operations to tcfsd via Unix domain socket gRPC.
+//!   Full fleet sync, NATS events, conflict resolution.
 
-use std::ffi::{CStr, CString};
+use std::ffi::CString;
 use std::os::raw::c_char;
-use std::panic::AssertUnwindSafe;
-use std::ptr;
-
-use base64::Engine;
-use secrecy::SecretString;
 
 /// Error codes returned by FFI functions.
 #[repr(C)]
@@ -73,617 +53,29 @@ pub struct TcfsFileItem {
     pub content_hash: *mut c_char,
 }
 
-/// Opaque provider handle wrapping a tokio runtime + OpenDAL operator.
-///
-/// Created via `tcfs_provider_new`, freed via `tcfs_provider_free`.
-pub struct TcfsProvider {
-    runtime: tokio::runtime::Runtime,
-    operator: opendal::Operator,
-    remote_prefix: String,
-    device_id: String,
-    /// Master key for E2EE (None = plaintext mode for backwards compatibility)
-    master_key: Option<tcfs_crypto::MasterKey>,
-}
+// ============================================================================
+// Direct backend (default): S3 via OpenDAL
+// ============================================================================
 
-/// Create a new provider from a JSON configuration string.
-///
-/// The JSON should contain:
-/// ```json
-/// {
-///   "s3_endpoint": "http://...",
-///   "s3_bucket": "tcfs",
-///   "s3_access": "...",
-///   "s3_secret": "...",
-///   "remote_prefix": "devices/mydevice"
-/// }
-/// ```
-///
-/// Returns a pointer to `TcfsProvider` on success, or null on failure.
-/// The caller must free the provider via `tcfs_provider_free`.
-///
-/// # Safety
-///
-/// `config_json` must be a valid null-terminated UTF-8 C string.
-#[no_mangle]
-pub unsafe extern "C" fn tcfs_provider_new(config_json: *const c_char) -> *mut TcfsProvider {
-    if config_json.is_null() {
-        return ptr::null_mut();
-    }
+#[cfg(feature = "direct")]
+mod direct;
 
-    let result = std::panic::catch_unwind(AssertUnwindSafe(|| {
-        let c_str = unsafe { CStr::from_ptr(config_json) };
-        let json_str = match c_str.to_str() {
-            Ok(s) => s,
-            Err(_) => return ptr::null_mut(),
-        };
+#[cfg(feature = "direct")]
+pub use direct::*;
 
-        let config: serde_json::Value = match serde_json::from_str(json_str) {
-            Ok(v) => v,
-            Err(_) => return ptr::null_mut(),
-        };
+// ============================================================================
+// gRPC backend: delegate to tcfsd daemon
+// ============================================================================
 
-        let endpoint = config["s3_endpoint"].as_str().unwrap_or_default();
-        let bucket = config["s3_bucket"].as_str().unwrap_or("tcfs");
-        let access = config["s3_access"].as_str().unwrap_or_default();
-        let secret = config["s3_secret"].as_str().unwrap_or_default();
-        let prefix = config["remote_prefix"]
-            .as_str()
-            .unwrap_or("default")
-            .to_string();
-        let device_id = config["device_id"]
-            .as_str()
-            .unwrap_or("unknown")
-            .to_string();
+#[cfg(feature = "grpc")]
+mod grpc_backend;
 
-        // Derive master key from encryption_passphrase if provided (enables E2EE)
-        let master_key = config["encryption_passphrase"]
-            .as_str()
-            .filter(|s| !s.is_empty())
-            .map(|passphrase| {
-                let salt_str = config["encryption_salt"]
-                    .as_str()
-                    .unwrap_or("tcfs-default-salt!");
-                let mut salt = [0u8; 16];
-                let salt_bytes = salt_str.as_bytes();
-                let copy_len = salt_bytes.len().min(16);
-                salt[..copy_len].copy_from_slice(&salt_bytes[..copy_len]);
+#[cfg(feature = "grpc")]
+pub use grpc_backend::*;
 
-                let params = tcfs_crypto::kdf::KdfParams {
-                    mem_cost_kib: config["argon2_mem_cost_kib"].as_u64().unwrap_or(65536) as u32,
-                    time_cost: config["argon2_time_cost"].as_u64().unwrap_or(3) as u32,
-                    parallelism: config["argon2_parallelism"].as_u64().unwrap_or(4) as u32,
-                };
-
-                tcfs_crypto::derive_master_key(
-                    &SecretString::from(passphrase.to_string()),
-                    &salt,
-                    &params,
-                )
-                .ok()
-            })
-            .flatten();
-
-        let runtime = match tokio::runtime::Runtime::new() {
-            Ok(rt) => rt,
-            Err(_) => return ptr::null_mut(),
-        };
-
-        let operator =
-            tcfs_storage::operator::build_operator(&tcfs_storage::operator::StorageConfig {
-                endpoint: endpoint.to_string(),
-                region: "us-east-1".to_string(),
-                bucket: bucket.to_string(),
-                access_key_id: access.to_string(),
-                secret_access_key: secret.to_string(),
-            });
-
-        let operator = match operator {
-            Ok(op) => op,
-            Err(_) => return ptr::null_mut(),
-        };
-
-        Box::into_raw(Box::new(TcfsProvider {
-            runtime,
-            operator,
-            remote_prefix: prefix,
-            device_id,
-            master_key,
-        }))
-    }));
-
-    result.unwrap_or(ptr::null_mut())
-}
-
-/// Enumerate files under a relative path within the remote prefix.
-///
-/// On success, writes an array of `TcfsFileItem` to `*out_items` and the
-/// count to `*out_count`. The caller must free the items via
-/// `tcfs_file_items_free`.
-///
-/// # Safety
-///
-/// - `provider` must be a valid pointer from `tcfs_provider_new`.
-/// - `path` must be a valid null-terminated UTF-8 C string (use "" for root).
-/// - `out_items` and `out_count` must be valid writable pointers.
-#[no_mangle]
-pub unsafe extern "C" fn tcfs_provider_enumerate(
-    provider: *mut TcfsProvider,
-    path: *const c_char,
-    out_items: *mut *mut TcfsFileItem,
-    out_count: *mut usize,
-) -> TcfsError {
-    if provider.is_null() || path.is_null() || out_items.is_null() || out_count.is_null() {
-        return TcfsError::TcfsErrorInvalidArg;
-    }
-
-    let result = std::panic::catch_unwind(AssertUnwindSafe(|| {
-        let prov = unsafe { &*provider };
-        let c_path = unsafe { CStr::from_ptr(path) };
-        let rel_path = match c_path.to_str() {
-            Ok(s) => s,
-            Err(_) => return TcfsError::TcfsErrorInvalidArg,
-        };
-
-        let prefix = format!(
-            "{}/index/{}",
-            prov.remote_prefix.trim_end_matches('/'),
-            rel_path.trim_start_matches('/')
-        );
-
-        let entries = match prov.runtime.block_on(prov.operator.list(&prefix)) {
-            Ok(e) => e,
-            Err(_) => return TcfsError::TcfsErrorStorage,
-        };
-
-        let mut items: Vec<TcfsFileItem> = Vec::new();
-        for entry in entries {
-            let entry_path = entry.path();
-            let name = entry_path
-                .strip_prefix(&prefix)
-                .unwrap_or(entry_path)
-                .trim_start_matches('/');
-
-            if name.is_empty() {
-                continue;
-            }
-
-            let is_dir = name.ends_with('/');
-            let display_name = name.trim_end_matches('/');
-
-            items.push(TcfsFileItem {
-                item_id: to_c_string(entry_path),
-                filename: to_c_string(display_name),
-                file_size: entry.metadata().content_length(),
-                modified_timestamp: 0,
-                is_directory: is_dir,
-                content_hash: to_c_string(""),
-            });
-        }
-
-        let count = items.len();
-        let boxed = items.into_boxed_slice();
-        let ptr = Box::into_raw(boxed) as *mut TcfsFileItem;
-
-        unsafe {
-            *out_items = ptr;
-            *out_count = count;
-        }
-
-        TcfsError::TcfsErrorNone
-    }));
-
-    result.unwrap_or(TcfsError::TcfsErrorInternal)
-}
-
-/// Fetch (hydrate) a file by its item ID to a local destination path.
-///
-/// Downloads chunks from S3, reassembles with integrity verification,
-/// and writes the result to `dest_path`.
-///
-/// # Safety
-///
-/// - `provider` must be a valid pointer from `tcfs_provider_new`.
-/// - `item_id` and `dest_path` must be valid null-terminated UTF-8 C strings.
-#[no_mangle]
-pub unsafe extern "C" fn tcfs_provider_fetch(
-    provider: *mut TcfsProvider,
-    item_id: *const c_char,
-    dest_path: *const c_char,
-) -> TcfsError {
-    if provider.is_null() || item_id.is_null() || dest_path.is_null() {
-        return TcfsError::TcfsErrorInvalidArg;
-    }
-
-    let result = std::panic::catch_unwind(AssertUnwindSafe(|| {
-        let prov = unsafe { &*provider };
-        let c_item = unsafe { CStr::from_ptr(item_id) };
-        let c_dest = unsafe { CStr::from_ptr(dest_path) };
-
-        let item_str = match c_item.to_str() {
-            Ok(s) => s,
-            Err(_) => return TcfsError::TcfsErrorInvalidArg,
-        };
-        let dest_str = match c_dest.to_str() {
-            Ok(s) => s,
-            Err(_) => return TcfsError::TcfsErrorInvalidArg,
-        };
-
-        let fetch_result = prov.runtime.block_on(async {
-            // Read the index entry to get manifest hash
-            let data = prov.operator.read(item_str).await?;
-            let bytes = data.to_bytes();
-            let text = String::from_utf8_lossy(&bytes);
-
-            let mut manifest_hash = String::new();
-            for line in text.lines() {
-                if let Some(val) = line.strip_prefix("manifest_hash=") {
-                    manifest_hash = val.to_string();
-                }
-            }
-
-            if manifest_hash.is_empty() {
-                anyhow::bail!("no manifest_hash in index entry");
-            }
-
-            let manifest_path = format!(
-                "{}/manifests/{}",
-                prov.remote_prefix.trim_end_matches('/'),
-                manifest_hash
-            );
-
-            let manifest_bytes = prov.operator.read(&manifest_path).await?;
-            let manifest =
-                tcfs_sync::manifest::SyncManifest::from_bytes(&manifest_bytes.to_bytes())?;
-
-            // Unwrap file key if E2EE manifest
-            let file_key = match (&prov.master_key, &manifest.encrypted_file_key) {
-                (Some(mk), Some(wrapped_b64)) => {
-                    let wrapped = base64::engine::general_purpose::STANDARD.decode(wrapped_b64)?;
-                    Some(tcfs_crypto::unwrap_key(mk, &wrapped)?)
-                }
-                (None, Some(_)) => {
-                    anyhow::bail!("file is encrypted but no master key configured");
-                }
-                _ => None,
-            };
-
-            // Reconstruct file_id for AAD verification (BLAKE3 of plaintext)
-            let file_id_bytes: [u8; 32] = tcfs_chunks::hash_from_hex(&manifest.file_hash)
-                .map(|h| *h.as_bytes())
-                .unwrap_or([0u8; 32]);
-
-            let mut assembled = Vec::new();
-            for (idx, hash) in manifest.chunk_hashes().iter().enumerate() {
-                let chunk_key = format!(
-                    "{}/chunks/{}",
-                    prov.remote_prefix.trim_end_matches('/'),
-                    hash
-                );
-                let chunk_data = prov.operator.read(&chunk_key).await?;
-                let chunk_bytes = chunk_data.to_bytes();
-
-                // BLAKE3 integrity verification (on ciphertext for E2EE, plaintext otherwise)
-                let actual = tcfs_chunks::hash_to_hex(&tcfs_chunks::hash_bytes(&chunk_bytes));
-                if actual != *hash {
-                    anyhow::bail!("chunk integrity failure: expected {}, got {}", hash, actual);
-                }
-
-                // Decrypt if E2EE
-                if let Some(ref fk) = file_key {
-                    let plaintext =
-                        tcfs_crypto::decrypt_chunk(fk, idx as u64, &file_id_bytes, &chunk_bytes)?;
-                    assembled.extend_from_slice(&plaintext);
-                } else {
-                    assembled.extend_from_slice(&chunk_bytes);
-                }
-            }
-
-            tokio::fs::write(dest_str, &assembled).await?;
-            Ok::<(), anyhow::Error>(())
-        });
-
-        match fetch_result {
-            Ok(()) => TcfsError::TcfsErrorNone,
-            Err(_) => TcfsError::TcfsErrorStorage,
-        }
-    }));
-
-    result.unwrap_or(TcfsError::TcfsErrorInternal)
-}
-
-/// Upload a local file to the remote prefix.
-///
-/// Chunks the file with FastCDC, hashes with BLAKE3, uploads chunks
-/// and manifest to S3.
-///
-/// # Safety
-///
-/// - `provider` must be a valid pointer from `tcfs_provider_new`.
-/// - `local_path` and `remote_rel` must be valid null-terminated UTF-8 C strings.
-#[no_mangle]
-pub unsafe extern "C" fn tcfs_provider_upload(
-    provider: *mut TcfsProvider,
-    local_path: *const c_char,
-    remote_rel: *const c_char,
-) -> TcfsError {
-    if provider.is_null() || local_path.is_null() || remote_rel.is_null() {
-        return TcfsError::TcfsErrorInvalidArg;
-    }
-
-    let result = std::panic::catch_unwind(AssertUnwindSafe(|| {
-        let prov = unsafe { &*provider };
-        let c_local = unsafe { CStr::from_ptr(local_path) };
-        let c_remote = unsafe { CStr::from_ptr(remote_rel) };
-
-        let local_str = match c_local.to_str() {
-            Ok(s) => s,
-            Err(_) => return TcfsError::TcfsErrorInvalidArg,
-        };
-        let remote_str = match c_remote.to_str() {
-            Ok(s) => s,
-            Err(_) => return TcfsError::TcfsErrorInvalidArg,
-        };
-
-        let upload_result = prov.runtime.block_on(async {
-            let data = tokio::fs::read(local_str).await?;
-            let file_hash = tcfs_chunks::hash_to_hex(&tcfs_chunks::hash_bytes(&data));
-
-            // Generate per-file encryption key if E2EE is enabled
-            let file_key = prov
-                .master_key
-                .as_ref()
-                .map(|_| tcfs_crypto::generate_file_key());
-            let file_id_bytes: [u8; 32] = {
-                let h = tcfs_chunks::hash_bytes(&data);
-                let mut arr = [0u8; 32];
-                arr.copy_from_slice(h.as_bytes());
-                arr
-            };
-
-            let chunks = tcfs_chunks::chunk_data(&data, tcfs_chunks::ChunkSizes::SMALL);
-            let mut chunk_hashes = Vec::new();
-
-            for (idx, chunk) in chunks.iter().enumerate() {
-                let chunk_bytes =
-                    &data[chunk.offset as usize..chunk.offset as usize + chunk.length];
-
-                // Encrypt chunk if E2EE is enabled, otherwise upload plaintext
-                let upload_bytes = if let Some(ref fk) = file_key {
-                    tcfs_crypto::encrypt_chunk(fk, idx as u64, &file_id_bytes, chunk_bytes)?
-                } else {
-                    chunk_bytes.to_vec()
-                };
-
-                let hash = tcfs_chunks::hash_to_hex(&tcfs_chunks::hash_bytes(&upload_bytes));
-                let chunk_key = format!(
-                    "{}/chunks/{}",
-                    prov.remote_prefix.trim_end_matches('/'),
-                    hash
-                );
-                prov.operator.write(&chunk_key, upload_bytes).await?;
-                chunk_hashes.push(hash);
-            }
-
-            // Build vclock: read existing manifest if present and merge
-            let mut vclock = tcfs_sync::conflict::VectorClock::new();
-            let existing_index_key = format!(
-                "{}/index/{}",
-                prov.remote_prefix.trim_end_matches('/'),
-                remote_str.trim_start_matches('/')
-            );
-            if let Ok(existing_data) = prov.operator.read(&existing_index_key).await {
-                let existing_bytes = existing_data.to_bytes();
-                let existing_text = String::from_utf8_lossy(&existing_bytes);
-                for line in existing_text.lines() {
-                    if let Some(hash) = line.strip_prefix("manifest_hash=") {
-                        let manifest_path = format!(
-                            "{}/manifests/{}",
-                            prov.remote_prefix.trim_end_matches('/'),
-                            hash
-                        );
-                        if let Ok(mb) = prov.operator.read(&manifest_path).await {
-                            if let Ok(existing_manifest) =
-                                tcfs_sync::manifest::SyncManifest::from_bytes(&mb.to_bytes())
-                            {
-                                vclock.merge(&existing_manifest.vclock);
-                            }
-                        }
-                    }
-                }
-            }
-            // Tick our device_id in the vclock
-            vclock.tick(&prov.device_id);
-
-            let written_at = std::time::SystemTime::now()
-                .duration_since(std::time::UNIX_EPOCH)
-                .unwrap_or_default()
-                .as_secs();
-
-            // Wrap file key with master key for storage in manifest
-            let encrypted_file_key = match (&prov.master_key, &file_key) {
-                (Some(mk), Some(fk)) => {
-                    let wrapped = tcfs_crypto::wrap_key(mk, fk)?;
-                    Some(base64::engine::general_purpose::STANDARD.encode(&wrapped))
-                }
-                _ => None,
-            };
-
-            let manifest = tcfs_sync::manifest::SyncManifest {
-                version: 2,
-                file_hash: file_hash.clone(),
-                file_size: data.len() as u64,
-                chunks: chunk_hashes,
-                vclock,
-                written_by: prov.device_id.clone(),
-                written_at,
-                rel_path: Some(remote_str.to_string()),
-                encrypted_file_key,
-            };
-
-            let manifest_json = serde_json::to_vec_pretty(&manifest)?;
-            let manifest_key = format!(
-                "{}/manifests/{}",
-                prov.remote_prefix.trim_end_matches('/'),
-                file_hash
-            );
-            prov.operator.write(&manifest_key, manifest_json).await?;
-
-            // Write index entry
-            let index_key = format!(
-                "{}/index/{}",
-                prov.remote_prefix.trim_end_matches('/'),
-                remote_str.trim_start_matches('/')
-            );
-            let index_entry = format!(
-                "manifest_hash={}\nsize={}\nchunks={}\n",
-                file_hash,
-                data.len(),
-                chunks.len()
-            );
-            prov.operator.write(&index_key, index_entry).await?;
-
-            Ok::<(), anyhow::Error>(())
-        });
-
-        match upload_result {
-            Ok(()) => TcfsError::TcfsErrorNone,
-            Err(_) => TcfsError::TcfsErrorStorage,
-        }
-    }));
-
-    result.unwrap_or(TcfsError::TcfsErrorInternal)
-}
-
-/// Delete a file or directory by its item ID (index path).
-///
-/// Performs a soft delete: removes the index entry and its manifest.
-/// Content-addressed chunks are left for GC (they may be shared).
-///
-/// # Safety
-///
-/// - `provider` must be a valid pointer from `tcfs_provider_new`.
-/// - `item_id` must be a valid null-terminated UTF-8 C string.
-#[no_mangle]
-pub unsafe extern "C" fn tcfs_provider_delete(
-    provider: *mut TcfsProvider,
-    item_id: *const c_char,
-) -> TcfsError {
-    if provider.is_null() || item_id.is_null() {
-        return TcfsError::TcfsErrorInvalidArg;
-    }
-
-    let result = std::panic::catch_unwind(AssertUnwindSafe(|| {
-        let prov = unsafe { &*provider };
-        let c_item = unsafe { CStr::from_ptr(item_id) };
-        let item_str = match c_item.to_str() {
-            Ok(s) => s,
-            Err(_) => return TcfsError::TcfsErrorInvalidArg,
-        };
-
-        let delete_result = prov.runtime.block_on(async {
-            // Read the index entry to get the manifest hash (for optional manifest cleanup)
-            let index_data = prov.operator.read(item_str).await;
-            if let Ok(data) = index_data {
-                let bytes = data.to_bytes();
-                let text = String::from_utf8_lossy(&bytes);
-                for line in text.lines() {
-                    if let Some(hash) = line.strip_prefix("manifest_hash=") {
-                        let manifest_path = format!(
-                            "{}/manifests/{}",
-                            prov.remote_prefix.trim_end_matches('/'),
-                            hash
-                        );
-                        // Best-effort manifest cleanup (content-addressed, may be shared)
-                        let _ = prov.operator.delete(&manifest_path).await;
-                    }
-                }
-            }
-
-            // Delete the index entry itself
-            prov.operator.delete(item_str).await?;
-            Ok::<(), anyhow::Error>(())
-        });
-
-        match delete_result {
-            Ok(()) => TcfsError::TcfsErrorNone,
-            Err(_) => TcfsError::TcfsErrorStorage,
-        }
-    }));
-
-    result.unwrap_or(TcfsError::TcfsErrorInternal)
-}
-
-/// Create a directory under the given parent path.
-///
-/// Writes a zero-byte marker at `{prefix}/index/{parent}/{dir_name}/`
-/// following the S3 directory convention.
-///
-/// # Safety
-///
-/// - `provider` must be a valid pointer from `tcfs_provider_new`.
-/// - `parent_path` and `dir_name` must be valid null-terminated UTF-8 C strings.
-#[no_mangle]
-pub unsafe extern "C" fn tcfs_provider_create_dir(
-    provider: *mut TcfsProvider,
-    parent_path: *const c_char,
-    dir_name: *const c_char,
-) -> TcfsError {
-    if provider.is_null() || parent_path.is_null() || dir_name.is_null() {
-        return TcfsError::TcfsErrorInvalidArg;
-    }
-
-    let result = std::panic::catch_unwind(AssertUnwindSafe(|| {
-        let prov = unsafe { &*provider };
-        let c_parent = unsafe { CStr::from_ptr(parent_path) };
-        let c_name = unsafe { CStr::from_ptr(dir_name) };
-
-        let parent_str = match c_parent.to_str() {
-            Ok(s) => s,
-            Err(_) => return TcfsError::TcfsErrorInvalidArg,
-        };
-        let name_str = match c_name.to_str() {
-            Ok(s) => s,
-            Err(_) => return TcfsError::TcfsErrorInvalidArg,
-        };
-
-        let dir_path = format!(
-            "{}/index/{}{}/",
-            prov.remote_prefix.trim_end_matches('/'),
-            if parent_str.is_empty() {
-                String::new()
-            } else {
-                format!("{}/", parent_str.trim_matches('/'))
-            },
-            name_str.trim_matches('/')
-        );
-
-        let create_result = prov
-            .runtime
-            .block_on(prov.operator.write(&dir_path, Vec::<u8>::new()));
-
-        match create_result {
-            Ok(_) => TcfsError::TcfsErrorNone,
-            Err(_) => TcfsError::TcfsErrorStorage,
-        }
-    }));
-
-    result.unwrap_or(TcfsError::TcfsErrorInternal)
-}
-
-/// Free a provider handle.
-///
-/// # Safety
-///
-/// `provider` must be a valid pointer from `tcfs_provider_new`, or null (no-op).
-/// Must not be called more than once for the same pointer.
-#[no_mangle]
-pub unsafe extern "C" fn tcfs_provider_free(provider: *mut TcfsProvider) {
-    if !provider.is_null() {
-        unsafe {
-            drop(Box::from_raw(provider));
-        }
-    }
-}
+// ============================================================================
+// Shared FFI helpers
+// ============================================================================
 
 /// Free an array of `TcfsFileItem` returned by `tcfs_provider_enumerate`.
 ///
@@ -704,7 +96,6 @@ pub unsafe extern "C" fn tcfs_file_items_free(items: *mut TcfsFileItem, count: u
             free_c_string(item.filename);
             free_c_string(item.content_hash);
         }
-        // Reconstruct the Box<[TcfsFileItem]> to drop it
         let _ = Box::from_raw(std::ptr::slice_from_raw_parts_mut(items, count));
     }
 }
@@ -718,8 +109,6 @@ pub unsafe extern "C" fn tcfs_file_items_free(items: *mut TcfsFileItem, count: u
 pub unsafe extern "C" fn tcfs_string_free(s: *mut c_char) {
     free_c_string(s);
 }
-
-// --- Internal helpers ---
 
 fn to_c_string(s: &str) -> *mut c_char {
     CString::new(s)
